feat: the file esm.mjs has been refactored
